requirements-linux.txt
joblib==0.16.0
grpcio==1.43.0
grpcio-tools==1.43.0
matplotlib==3.3.1
pytest==6.2.1
pandas==1.1.0
Pillow==8.1.0
protobuf==3.14.0
tensorflow==2.4.0
numpy==1.19.2
scikit-image==0.19.2        # for noise cancellation